More std::size_t refactoring.

+++ include/skimdb/spir/skimdb_spir_definitions.h
@@ -26,10 +26,10 @@ struct spirdb_parameters {
   std::uint64_t n; // LWE dimension
   double sigma;    // LWE error distribution stddev
 
-  std::uint32_t log_p; // plaintext modulus
-  std::uint32_t log_q; // ciphertext modulus
+  std::size_t log_p; // plaintext modulus
+  std::size_t log_q; // ciphertext modulus
 
-  std::uint32_t batch_size; // number of qu vectors which can be processed in one batch
+  std::size_t batch_size; // number of qu vectors which can be processed in one batch
 
   std::uint32_t block_len;  // bytes of plaintext data we can pack into one block
   std::uint64_t rle_blocks; // number of blocks needed per RLE encoding
